+++ empire/test/test_agent_task_api.py
@@ -1,6 +1,18 @@
 from textwrap import dedent
+from types import SimpleNamespace
 
 import pytest
+
+from empire.server.core.exceptions import (
+    ModuleExecutionException,
+    ModuleValidationException,
+)
+from empire.server.core.module_models import (
+    EmpireModule,
+    EmpireModuleAdvanced,
+    LanguageEnum,
+)
+from empire.server.utils.module_util import handle_error_message
 
 
 @pytest.fixture(scope="module", autouse=True)
@@ -176,6 +188,82 @@ def agent_task(client, admin_auth_header, agent):
 
     # No need to delete the task, it will be deleted when the agent is deleted
     # After the test.
+
+
+def raise_exception_wrapper(exception):
+    def raise_exception(*args, **kwargs):
+        raise exception
+
+    return raise_exception
+
+
+def return_handle_error_message_wrapper(message):
+    def return_handle_error_message(*args, **kwargs):
+        return handle_error_message(message)
+
+    return return_handle_error_message
+
+
+@pytest.fixture(scope="module", autouse=True)
+def module_with_validation_exception(main):
+    module_name = "this_module_has_a_validation_exception"
+    main.modulesv2.modules[module_name] = EmpireModule(
+        id=module_name,
+        name=module_name,
+        language=LanguageEnum.powershell,
+        advanced=EmpireModuleAdvanced(
+            custom_generate=True,
+            generate_class=SimpleNamespace(
+                generate=raise_exception_wrapper(ModuleValidationException(module_name))
+            ),
+        ),
+    )
+
+    yield
+
+    del main.modulesv2.modules[module_name]
+
+
+@pytest.fixture(scope="module", autouse=True)
+def module_with_execution_exception(main):
+    module_name = "this_module_has_an_execution_exception"
+    main.modulesv2.modules[module_name] = EmpireModule(
+        id=module_name,
+        name=module_name,
+        language=LanguageEnum.powershell,
+        advanced=EmpireModuleAdvanced(
+            custom_generate=True,
+            generate_class=SimpleNamespace(
+                generate=raise_exception_wrapper(ModuleExecutionException(module_name))
+            ),
+        ),
+    )
+
+    yield
+
+    del main.modulesv2.modules[module_name]
+
+
+@pytest.fixture(scope="module", autouse=True)
+def module_with_legacy_handle_error_message(main):
+    module_name = "this_module_uses_legacy_handle_error_message"
+    main.modulesv2.modules[module_name] = EmpireModule(
+        id=module_name,
+        name=module_name,
+        language=LanguageEnum.powershell,
+        advanced=EmpireModuleAdvanced(
+            custom_generate=True,
+            generate_class=SimpleNamespace(
+                generate=return_handle_error_message_wrapper(
+                    module_name + ": this is the error"
+                )
+            ),
+        ),
+    )
+
+    yield
+
+    del main.modulesv2.modules[module_name]
 
 
 def test_create_task_shell_agent_not_found(client, admin_auth_header):
@@ -442,6 +530,60 @@ def test_create_task_module_ignore_admin_check(
 
     assert response.status_code == 201
     assert response.json()["id"] > 0
+
+
+def test_create_task_module_validation_exception(
+    client, admin_auth_header, agent_low_integrity
+):
+    response = client.post(
+        f"/api/v2/agents/{agent_low_integrity.session_id}/tasks/module",
+        headers=admin_auth_header,
+        json={
+            "module_id": "this_module_uses_legacy_handle_error_message",
+            "ignore_admin_check": True,
+            "options": {},
+        },
+    )
+
+    assert response.status_code == 400
+    assert (
+        response.json()["detail"]
+        == "this_module_uses_legacy_handle_error_message: this is the error"
+    )
+
+
+def test_create_task_module_execution_exception(
+    client, admin_auth_header, agent_low_integrity
+):
+    response = client.post(
+        f"/api/v2/agents/{agent_low_integrity.session_id}/tasks/module",
+        headers=admin_auth_header,
+        json={
+            "module_id": "this_module_has_an_execution_exception",
+            "ignore_admin_check": True,
+            "options": {},
+        },
+    )
+
+    assert response.status_code == 500
+    assert response.json()["detail"] == "this_module_has_an_execution_exception"
+
+
+def test_create_task_handle_error_message(
+    client, admin_auth_header, agent_low_integrity
+):
+    response = client.post(
+        f"/api/v2/agents/{agent_low_integrity.session_id}/tasks/module",
+        headers=admin_auth_header,
+        json={
+            "module_id": "this_module_has_an_execution_exception",
+            "ignore_admin_check": True,
+            "options": {},
+        },
+    )
+
+    assert response.status_code == 500
+    assert response.json()["detail"] == "this_module_has_an_execution_exception"
 
 
 def test_create_task_upload_file_not_found(client, admin_auth_header, agent):